% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music.R
\name{now_playing_music}
\alias{now_playing_music}
\title{JS Music Playing}
\usage{
now_playing_music(message, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{message}{A keyword representing your current mood}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Required by htmlwidgets. Useless.}
}
\description{
Music user interface using JS by 'https://spinitron.github.io/v2-web-integration/widget.html'.
Support random music playing in both R-markdwon file and HTML file.
}
