% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music.R
\name{netease_music}
\alias{netease_music}
\title{Netease Music}
\usage{
netease_music(query = "\\u4ee3\\u7801", show_in_viewer = TRUE, select = 1)
}
\arguments{
\item{query}{Key word that neads to be searched.}

\item{show_in_viewer}{Binary variable with 1 representing the music interface need to be shown in RStudio Viewer.}

\item{select}{Number of songs to be included.}
}
\description{
Netease interface in R-markdown file and HTML file. If run in R-markdown chunk, interface
will be displayed in RStudio viewer. If compiled to HTML, interface will be embedded under chunk.
Support key word searching.
}
